dom/serviceworkers/test/test_notificationclose.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1265841
-->
<head>
  <title>Bug 1265841 - Test ServiceWorkerGlobalScope.notificationclose event.</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/dom/notification/test/mochitest/MockAlertsService.js"></script>
  <script type="text/javascript" src="/tests/dom/notification/test/mochitest/NotificationTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1265841">Bug 1265841</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>
<pre id="test">
</pre>
<script src="utils.js"></script>
<script type="text/javascript">
  SimpleTest.requestFlakyTimeout("Mock alert service dispatches show, click, and close events.");

  function testFrame(src) {
    var iframe = document.createElement("iframe");
    iframe.src = src;
    window.callback = async function(data) {
      window.callback = null;
      document.body.removeChild(iframe);
      iframe = null;
      ok(data.result, "Got notificationclose event with correct data.");
      ok(!data.windowOpened,
         "Shouldn't allow to openWindow in notificationclose");
      await registration.unregister();
      SimpleTest.finish();
    };
    document.body.appendChild(iframe);
  }

  var registration;

  async function runTest() {
    try {
      await NotificationTest.allowNotifications();
      await MockAlertsService.register();
      await MockAlertsService.enableAutoClick();
      registration = await navigator.serviceWorker.register(
        "notificationclose.js",
        { scope: "notificationclose.html" }
      );
      await waitForState(registration.installing, 'activated');
    } catch (e) {
      ok(false, "registration should have passed!");
      SimpleTest.finish();
    }
    testFrame('notificationclose.html');
  };

  SimpleTest.waitForExplicitFinish();
  SpecialPowers.pushPrefEnv({"set": [
    ["dom.serviceWorkers.exemptFromPerDomainMax", true],
    ["dom.serviceWorkers.enabled", true],
    ["dom.serviceWorkers.testing.enabled", true],
  ]}, runTest);
</script>
</body>
</html>
